Refactoring usage of frequency and encoderinput fields
Renamed "EncoderInput" to "InputType" since it's only set to
eTelevision, eRadio or eExternalInput. It's used for general
decisions depending on the type of the input.
On the contrary "CurrentInput" is holding the actual input number
of the device.

Naming convention:
Properties of current set channel should start with "Current".
Properties of the channel to be set should start with "new".